# api/cmd/kubermatic-api/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Kubermatic API\n\nThis describes possible operations which can be made against the Kubermatic API.",
    "title": "Kubermatic API.",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "version": "2.2.3"
  },
  "host": "cloud.kubermatic.io",
  "paths": {
...
    "/api/v1/projects/{project_id}/dc/{dc}/clusters/{cluster_id}/kubeconfig": {
      "get": {
        "produces": [
          "application/yaml"
        ],
        "tags": [
          "project"
        ],
        "summary": "Gets the kubeconfig for the specified cluster.",
        "operationId": "newGetClusterKubeconfig",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ProjectID",
            "name": "project_id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "DC",
            "name": "dc",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "ClusterID",
            "name": "cluster_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Kubeconfig",
            "schema": {
              "$ref": "#/definitions/Kubeconfig"
            }
          },
          "401": {
            "$ref": "#/responses/empty"
          },
          "403": {
            "$ref": "#/responses/empty"
          },
          "default": {
            "$ref": "#/responses/errorResponse"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AWSCloudSpec": {
      "type": "object",
      "title": "AWSCloudSpec specifies access data to Amazon Web Services.",
      "properties": {
        "accessKeyId": {
          "type": "string",
          "x-go-name": "AccessKeyID"
        },
        "availabilityZone": {
          "type": "string",
          "x-go-name": "AvailabilityZone"
        },
        "instanceProfileName": {
          "type": "string",
          "x-go-name": "InstanceProfileName"
        },
        "roleName": {
          "type": "string",
          "x-go-name": "RoleName"
        },
        "routeTableId": {
          "type": "string",
          "x-go-name": "RouteTableID"
        },
        "secretAccessKey": {
          "type": "string",
          "x-go-name": "SecretAccessKey"
        },
        "securityGroupID": {
          "type": "string",
          "x-go-name": "SecurityGroupID"
        },
        "subnetId": {
          "type": "string",
          "x-go-name": "SubnetID"
        },
        "vpcId": {
          "type": "string",
          "x-go-name": "VPCID"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
    },
...
    "AuthInfo": {
      "type": "object",
      "title": "AuthInfo contains information that describes identity information.  This is use to tell the kubernetes cluster who you are.",
      "properties": {
        "as": {
          "description": "Impersonate is the username to imperonate.  The name matches the flag.\n+optional",
          "type": "string",
          "x-go-name": "Impersonate"
        },
        "as-groups": {
          "description": "ImpersonateGroups is the groups to imperonate.\n+optional",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "ImpersonateGroups"
        },
        "as-user-extra": {
          "description": "ImpersonateUserExtra contains additional information for impersonated user.\n+optional",
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "x-go-name": "ImpersonateUserExtra"
        },
        "auth-provider": {
          "$ref": "#/definitions/AuthProviderConfig"
        },
        "client-certificate": {
          "description": "ClientCertificate is the path to a client cert file for TLS.\n+optional",
          "type": "string",
          "x-go-name": "ClientCertificate"
        },
        "client-certificate-data": {
          "description": "ClientCertificateData contains PEM-encoded data from a client cert file for TLS. Overrides ClientCertificate\n+optional",
          "type": "array",
          "items": {
            "type": "integer",
            "format": "uint8"
          },
          "x-go-name": "ClientCertificateData"
        },
        "client-key": {
          "description": "ClientKey is the path to a client key file for TLS.\n+optional",
          "type": "string",
          "x-go-name": "ClientKey"
        },
        "client-key-data": {
          "description": "ClientKeyData contains PEM-encoded data from a client key file for TLS. Overrides ClientKey\n+optional",
          "type": "array",
          "items": {
            "type": "integer",
            "format": "uint8"
          },
          "x-go-name": "ClientKeyData"
        },
        "exec": {
          "$ref": "#/definitions/ExecConfig"
        },
        "extensions": {
          "description": "Extensions holds additional information. This is useful for extenders so that reads and writes don't clobber unknown fields\n+optional",
          "type": "array",
          "items": {
            "$ref": "#/definitions/NamedExtension"
          },
          "x-go-name": "Extensions"
        },
        "password": {
          "description": "Password is the password for basic authentication to the kubernetes cluster.\n+optional",
          "type": "string",
          "x-go-name": "Password"
        },
        "token": {
          "description": "Token is the bearer token for authentication to the kubernetes cluster.\n+optional",
          "type": "string",
          "x-go-name": "Token"
        },
        "tokenFile": {
          "description": "TokenFile is a pointer to a file that contains a bearer token (as described above).  If both Token and TokenFile are present, Token takes precedence.\n+optional",
          "type": "string",
          "x-go-name": "TokenFile"
        },
        "username": {
          "description": "Username is the username for basic authentication to the kubernetes cluster.\n+optional",
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/vendor/k8s.io/client-go/tools/clientcmd/api/v1"
    },
    "AuthProviderConfig": {
      "type": "object",
      "title": "AuthProviderConfig holds the configuration for a specified auth provider.",
      "properties": {
        "config": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Config"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/vendor/k8s.io/client-go/tools/clientcmd/api/v1"
    },
    "AzureCloudSpec": {
      "type": "object",
      "title": "AzureCloudSpec specifies acceess credentials to Azure cloud.",
      "properties": {
        "availabilitySet": {
          "type": "string",
          "x-go-name": "AvailabilitySet"
        },
        "clientID": {
          "type": "string",
          "x-go-name": "ClientID"
        },
        "clientSecret": {
          "type": "string",
          "x-go-name": "ClientSecret"
        },
        "resourceGroup": {
          "type": "string",
          "x-go-name": "ResourceGroup"
        },
        "routeTable": {
          "type": "string",
          "x-go-name": "RouteTableName"
        },
        "securityGroup": {
          "type": "string",
          "x-go-name": "SecurityGroup"
        },
        "subnet": {
          "type": "string",
          "x-go-name": "SubnetName"
        },
        "subscriptionID": {
          "type": "string",
          "x-go-name": "SubscriptionID"
        },
        "tenantID": {
          "type": "string",
          "x-go-name": "TenantID"
        },
        "vnet": {
          "type": "string",
          "x-go-name": "VNetName"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
    },
...
    "BringYourOwnCloudSpec": {
      "type": "object",
      "title": "BringYourOwnCloudSpec specifies access data for a bring your own cluster.",
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
    },
...
    "CloudSpec": {
      "type": "object",
      "title": "CloudSpec mutually stores access data to a cloud provider.",
      "properties": {
        "aws": {
          "$ref": "#/definitions/AWSCloudSpec"
        },
        "azure": {
          "$ref": "#/definitions/AzureCloudSpec"
        },
        "bringyourown": {
          "$ref": "#/definitions/BringYourOwnCloudSpec"
        },
        "dc": {
          "description": "DatacenterName where the users 'cloud' lives in.",
          "type": "string",
          "x-go-name": "DatacenterName"
        },
        "digitalocean": {
          "$ref": "#/definitions/DigitaloceanCloudSpec"
        },
        "fake": {
          "$ref": "#/definitions/FakeCloudSpec"
        },
        "hetzner": {
          "$ref": "#/definitions/HetznerCloudSpec"
        },
        "openstack": {
          "$ref": "#/definitions/OpenstackCloudSpec"
        },
        "vsphere": {
          "$ref": "#/definitions/VSphereCloudSpec"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
    },
    "Cluster": {
      "description": "NewCluster defines the cluster resource",
      "type": "object",
      "properties": {
        "creationTimestamp": {
          "description": "CreationTimestamp is a timestamp representing the server time when this object was created.",
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreationTimestamp"
        },
        "deletionTimestamp": {
          "description": "DeletionTimestamp is a timestamp representing the server time when this object was deleted.",
          "type": "string",
          "format": "date-time",
          "x-go-name": "DeletionTimestamp"
        },
        "id": {
          "description": "ID unique value that identifies the resource generated by the server",
          "type": "string",
          "x-go-name": "ID"
        },
        "name": {
          "description": "Name represents human readable name for the resource",
          "type": "string",
          "x-go-name": "Name"
        },
        "spec": {
          "$ref": "#/definitions/NewClusterSpec"
        },
        "status": {
          "$ref": "#/definitions/NewClusterStatus"
        }
      },
      "x-go-name": "NewCluster",
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/api/v1"
    },
...
    "Context": {
      "description": "Context is a tuple of references to a cluster (how do I communicate with a kubernetes cluster), a user (how do I identify myself), and a namespace (what subset of resources do I want to work with)",
      "type": "object",
      "properties": {
        "cluster": {
          "description": "Cluster is the name of the cluster for this context",
          "type": "string",
          "x-go-name": "Cluster"
        },
        "extensions": {
          "description": "Extensions holds additional information. This is useful for extenders so that reads and writes don't clobber unknown fields\n+optional",
          "type": "array",
          "items": {
            "$ref": "#/definitions/NamedExtension"
          },
          "x-go-name": "Extensions"
        },
        "namespace": {
          "description": "Namespace is the default namespace to use on unspecified requests\n+optional",
          "type": "string",
          "x-go-name": "Namespace"
        },
        "user": {
          "description": "AuthInfo is the name of the authInfo for this context",
          "type": "string",
          "x-go-name": "AuthInfo"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/vendor/k8s.io/client-go/tools/clientcmd/api/v1"
    },
...
    "DigitaloceanCloudSpec": {
      "type": "object",
      "title": "DigitaloceanCloudSpec specifies access data to DigitalOcean.",
      "properties": {
        "token": {
          "type": "string",
          "x-go-name": "Token"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
    },
...
    "ErrorDetails": {
      "description": "ErrorDetails contains details about the error",
      "type": "object",
      "required": [
        "code",
        "message"
      ],
      "properties": {
        "code": {
          "description": "The error code",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ErrorCode"
        },
        "message": {
          "description": "The error message",
          "type": "string",
          "x-go-name": "ErrorMessage"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/handler"
    },
...
    "ExecConfig": {
      "description": "See the client.authentiction.k8s.io API group for specifications of the exact input\nand output format",
      "type": "object",
      "title": "ExecConfig specifies a command to provide client credentials. The command is exec'd\nand outputs structured stdout holding credentials.",
      "properties": {
        "apiVersion": {
          "description": "Preferred input version of the ExecInfo. The returned ExecCredentials MUST use\nthe same encoding version as the input.",
          "type": "string",
          "x-go-name": "APIVersion"
        },
        "args": {
          "description": "Arguments to pass to the command when executing it.\n+optional",
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Args"
        },
        "command": {
          "description": "Command to execute.",
          "type": "string",
          "x-go-name": "Command"
        },
        "env": {
          "description": "Env defines additional environment variables to expose to the process. These\nare unioned with the host's environment, as well as variables client-go uses\nto pass argument to the plugin.\n+optional",
          "type": "array",
          "items": {
            "$ref": "#/definitions/ExecEnvVar"
          },
          "x-go-name": "Env"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/vendor/k8s.io/client-go/tools/clientcmd/api/v1"
    },
    "ExecEnvVar": {
      "description": "ExecEnvVar is used for setting environment variables when executing an exec-based\ncredential plugin.",
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "value": {
          "type": "string",
          "x-go-name": "Value"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/vendor/k8s.io/client-go/tools/clientcmd/api/v1"
    },
    "FakeCloudSpec": {
      "type": "object",
      "title": "FakeCloudSpec specifies access data for a fake cloud.",
      "properties": {
        "token": {
          "type": "string",
          "x-go-name": "Token"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
    },
    "HetznerCloudSpec": {
      "type": "object",
      "title": "HetznerCloudSpec specifies access data to hetzner cloud.",
      "properties": {
        "token": {
          "type": "string",
          "x-go-name": "Token"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
    },
...
    "Kubeconfig": {
      "description": "Kubeconfig is a clusters kubeconfig",
      "type": "object",
      "properties": {
        "apiVersion": {
          "description": "Legacy field from pkg/api/types.go TypeMeta.\nTODO(jlowdermilk): remove this after eliminating downstream dependencies.\n+optional",
          "type": "string",
          "x-go-name": "APIVersion"
        },
        "clusters": {
          "description": "Clusters is a map of referencable names to cluster configs",
          "type": "array",
          "items": {
            "$ref": "#/definitions/NamedCluster"
          },
          "x-go-name": "Clusters"
        },
        "contexts": {
          "description": "Contexts is a map of referencable names to context configs",
          "type": "array",
          "items": {
            "$ref": "#/definitions/NamedContext"
          },
          "x-go-name": "Contexts"
        },
        "current-context": {
          "description": "CurrentContext is the name of the context that you would like to use by default",
          "type": "string",
          "x-go-name": "CurrentContext"
        },
        "extensions": {
          "description": "Extensions holds additional information. This is useful for extenders so that reads and writes don't clobber unknown fields\n+optional",
          "type": "array",
          "items": {
            "$ref": "#/definitions/NamedExtension"
          },
          "x-go-name": "Extensions"
        },
        "kind": {
          "description": "Legacy field from pkg/api/types.go TypeMeta.\nTODO(jlowdermilk): remove this after eliminating downstream dependencies.\n+optional",
          "type": "string",
          "x-go-name": "Kind"
        },
        "preferences": {
          "$ref": "#/definitions/Preferences"
        },
        "users": {
          "description": "AuthInfos is a map of referencable names to user configs",
          "type": "array",
          "items": {
            "$ref": "#/definitions/NamedAuthInfo"
          },
          "x-go-name": "AuthInfos"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/api/v1"
    },
    "MachineNetworkingConfig": {
      "type": "object",
      "title": "MachineNetworkingConfig specifies the networking parameters used for IPAM.",
      "properties": {
        "cidr": {
          "type": "string",
          "x-go-name": "CIDR"
        },
        "dnsServers": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "DNSServers"
        },
        "gateway": {
          "type": "string",
          "x-go-name": "Gateway"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
    },
...
    "NamedAuthInfo": {
      "description": "NamedAuthInfo relates nicknames to auth information",
      "type": "object",
      "properties": {
        "name": {
          "description": "Name is the nickname for this AuthInfo",
          "type": "string",
          "x-go-name": "Name"
        },
        "user": {
          "$ref": "#/definitions/AuthInfo"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/vendor/k8s.io/client-go/tools/clientcmd/api/v1"
    },
    "NamedCluster": {
      "description": "NamedCluster relates nicknames to cluster information",
      "type": "object",
      "properties": {
        "cluster": {
          "$ref": "#/definitions/Cluster"
        },
        "name": {
          "description": "Name is the nickname for this Cluster",
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/vendor/k8s.io/client-go/tools/clientcmd/api/v1"
    },
    "NamedContext": {
      "description": "NamedContext relates nicknames to context information",
      "type": "object",
      "properties": {
        "context": {
          "$ref": "#/definitions/Context"
        },
        "name": {
          "description": "Name is the nickname for this Context",
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/vendor/k8s.io/client-go/tools/clientcmd/api/v1"
    },
    "NamedExtension": {
      "description": "NamedExtension relates nicknames to extension information",
      "type": "object",
      "properties": {
        "extension": {
          "$ref": "#/definitions/RawExtension"
        },
        "name": {
          "description": "Name is the nickname for this Extension",
          "type": "string",
          "x-go-name": "Name"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/vendor/k8s.io/client-go/tools/clientcmd/api/v1"
    },
...
    "NewClusterSpec": {
      "description": "NewClusterSpec defines the cluster specification",
      "type": "object",
      "properties": {
        "cloud": {
          "$ref": "#/definitions/CloudSpec"
        },
        "machineNetworks": {
          "description": "MachineNetworks optionally specifies the parameters for IPAM.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/MachineNetworkingConfig"
          },
          "x-go-name": "MachineNetworks"
        },
        "version": {
          "description": "Version desired version of the kubernetes master components",
          "type": "string",
          "x-go-name": "Version"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/api/v1"
    },
    "NewClusterStatus": {
      "description": "NewClusterStatus defines the cluster status",
      "type": "object",
      "properties": {
        "url": {
          "description": "URL specifies the address at which the cluster is available",
          "type": "string",
          "x-go-name": "URL"
        },
        "version": {
          "description": "Version actual version of the kubernetes master components",
          "type": "string",
          "x-go-name": "Version"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/api/v1"
    },
...
    "OpenstackCloudSpec": {
      "type": "object",
      "title": "OpenstackCloudSpec specifies access data to an openstack cloud.",
      "properties": {
        "domain": {
          "type": "string",
          "x-go-name": "Domain"
        },
        "floatingIpPool": {
          "type": "string",
          "x-go-name": "FloatingIPPool"
        },
        "network": {
          "type": "string",
          "x-go-name": "Network"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "routerID": {
          "type": "string",
          "x-go-name": "RouterID"
        },
        "securityGroups": {
          "type": "string",
          "x-go-name": "SecurityGroups"
        },
        "subnetID": {
          "type": "string",
          "x-go-name": "SubnetID"
        },
        "tenant": {
          "type": "string",
          "x-go-name": "Tenant"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
    },
...
    "Preferences": {
      "type": "object",
      "properties": {
        "colors": {
          "description": "+optional",
          "type": "boolean",
          "x-go-name": "Colors"
        },
        "extensions": {
          "description": "Extensions holds additional information. This is useful for extenders so that reads and writes don't clobber unknown fields\n+optional",
          "type": "array",
          "items": {
            "$ref": "#/definitions/NamedExtension"
          },
          "x-go-name": "Extensions"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/vendor/k8s.io/client-go/tools/clientcmd/api/v1"
    },
...
    "RawExtension": {
      "description": "To use this, make a field which has RawExtension as its type in your external, versioned\nstruct, and Object in your internal struct. You also need to register your\nvarious plugin types.\n\nInternal package:\ntype MyAPIObject struct {\nruntime.TypeMeta `json:\",inline\"`\nMyPlugin runtime.Object `json:\"myPlugin\"`\n}\ntype PluginA struct {\nAOption string `json:\"aOption\"`\n}\n\nExternal package:\ntype MyAPIObject struct {\nruntime.TypeMeta `json:\",inline\"`\nMyPlugin runtime.RawExtension `json:\"myPlugin\"`\n}\ntype PluginA struct {\nAOption string `json:\"aOption\"`\n}\n\nOn the wire, the JSON will look something like this:\n{\n\"kind\":\"MyAPIObject\",\n\"apiVersion\":\"v1\",\n\"myPlugin\": {\n\"kind\":\"PluginA\",\n\"aOption\":\"foo\",\n},\n}\n\nSo what happens? Decode first uses json or yaml to unmarshal the serialized data into\nyour external MyAPIObject. That causes the raw JSON to be stored, but not unpacked.\nThe next step is to copy (using pkg/conversion) into the internal struct. The runtime\npackage's DefaultScheme has conversion functions installed which will unpack the\nJSON stored in RawExtension, turning it into the correct object type, and storing it\nin the Object. (TODO: In the case where the object is of an unknown type, a\nruntime.Unknown object will be created and stored.)\n\n+k8s:deepcopy-gen=true\n+protobuf=true\n+k8s:openapi-gen=true",
      "type": "object",
      "title": "RawExtension is used to hold extensions in external versions.",
      "properties": {
        "Raw": {
          "description": "Raw is the underlying serialization of this object.\n\nTODO: Determine how to detect ContentType and ContentEncoding of 'Raw' data.",
          "type": "array",
          "items": {
            "type": "integer",
            "format": "uint8"
          }
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/vendor/k8s.io/apimachinery/pkg/runtime"
    },
...
    "VSphereCloudSpec": {
      "type": "object",
      "title": "VSphereCloudSpec specifies access data to VSphere cloud.",
      "properties": {
        "infraManagementUser": {
          "$ref": "#/definitions/VSphereCredentials"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        },
        "vmNetName": {
          "type": "string",
          "x-go-name": "VMNetName"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
    },
    "VSphereCredentials": {
      "description": "VSphereCredentials credentials represents a credential for accessing vSphere",
      "type": "object",
      "properties": {
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "github.com/kubermatic/kubermatic/api/pkg/crd/kubermatic/v1"
    },
...
  },
  "responses": {
    "empty": {
      "description": "EmptyResponse is a empty response"
    },
    "errorResponse": {
      "description": "ErrorResponse is the default representation of an error",
      "schema": {
        "$ref": "#/definitions/ErrorDetails"
      }
    }
  }
}